create basic App, Display, ButtonsPanel and Button stateless functions components
