+++ src/components/App.js
@@ -1,0 +1,13 @@
+import Display from './Display';
+import Buttonspanel from './ButtonsPanel';
+
+function App() {
+  return (
+    <div className="App">
+      <Display />
+      <Buttonspanel />
+    </div>
+  );
+}
+
+export default App;
