+++ src/components/App.test.js
@@ -1,0 +1,6 @@
+import { render } from '@testing-library/react';
+import App from './App';
+
+test('renders learn react link', () => {
+  render(<App />);
+});
