+++ src/components/Button.js
@@ -1,0 +1,7 @@
+const Button = () => (
+  <div>
+    Button
+  </div>
+);
+
+export default Button;
