+++ src/components/ButtonsPanel.js
@@ -1,0 +1,10 @@
+import Button from './Button';
+
+const ButtonsPanel = () => (
+  <div>
+    ButtonsPanel
+    <Button />
+  </div>
+);
+
+export default ButtonsPanel;
