+++ src/components/Display.js
@@ -1,0 +1,7 @@
+const Display = () => (
+  <div>
+    Display
+  </div>
+);
+
+export default Display;
